Massive refactor, not production ready
