+++ Cargo.toml
@@ -17,14 +17,15 @@ log = "0.4.22"
 needletail = "^0.6.3"
 rayon = "1.10.0"
 regex = "1.10.6"
-serde = { version = "1.0.203", features = ["derive"] }
-serde_json = "1.0.118"
 spoa = { git = "https://github.com/olliecheng/spoa-rs" }
 thiserror = "1.0.64"
 predicates = "3.1.2"
 indoc = "2.0.5"
 rkyv = { version = "0.8.10", features = ["alloc", "bytecheck", "std", "indexmap-2", "smallvec-1"] }
 smallvec = "1.15.0"
+memmap = "0.7.0"
+bio = "2.2.0"
+itertools = "0.14.0"
 
 [[bin]]
 name = "nailpolish"
